Remove Custom Hero

+++ Datapack/authserver/sql/install/accounts.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS `accounts` (
   `login` varchar(32) NOT NULL,
   `password` varchar(255) NOT NULL,
-  `last_access` int(11) NOT NULL DEFAULT '0',
+  `last_access` DECIMAL(20, 0) NOT NULL DEFAULT '0',
   `access_level` int(11) NOT NULL DEFAULT '0',
   `last_ip` varchar(15) DEFAULT NULL,
   `last_server` int(11) NOT NULL DEFAULT '0',
